Update synology_SMART_control.sh
added re-scan and re-save of the disk data to ```disk_scan_status.txt``` whenever a manual test start or manual test cancellation is performed by the user

the reason for this is:
when the manual command was submitted in the web-interface, the test was either started or canceled correctly upon the next script execution. However because the ```disk_scan_status.txt``` file contents are saved only at the beginning of the script, PRIOR to the tests being started or cancelled, the web-interface would NOT update until ANOTHER entire waiting period expired and the script was executed again. 

confirmed to now allow the web interface to display the correct disk status as soon as the manual start/cancel operation is executed by the script.