Update synology_SMART_control.sh
added logo to web interface addressing issue #30
added text color formatting to web interface
updated web-interface to not cause a form re submission when refreshing the page addressing issue #21
added script version detail to web-interface
updated logic to web-interface to use system Host-Name if "System Name" in web interface is left blank addressing issue #28
updated script logic to ensure next scan time schedule is recalculated and saved if the time to perform a scan occurs, BUT either a manaul test is active or a BTRFS/RAID scrub is active. this ensures the schedule is not thrown off. addressed issue #25
updated script so that the web-interface data is always saved at the beginning of the file and removed the saving from the beginning of the file. this ensures the web-interface status always reflects the latest disk scan status
updated and added images to the read me file showing latest web-interface layout and colors addressing issue #27
removed "extended" from the email notifications since both short and long tests can be started and cancelled.